requirements.txt
certifi==2015.4.28
Django==1.8.2
django-haystack==2.3.1
elasticsearch==1.5.0
elasticstack==0.1.1
pyelasticsearch==1.3
simplejson==3.7.3
six==1.9.0
urllib3==1.10.4
